pyproject.toml
[tool.poetry]
name = "enterprise-mcp-mvp-demo-ui"
version = "0.1.0"
description = ""
authors = ["fisseha-estifanos <fisseha.137@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
streamlit = "^1.41.1"
watchdog = "^6.0.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
